style: 💄 section order changes in form editor card
